Add TODO comment and remove unused files
