--- carma-messenger-core/j2735_convertor/src/main_ros2.cpp
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2022 LEIDOS.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-#include <rclcpp/rclcpp.hpp>
-#include "j2735_convertor/j2735_convertor_node.hpp"
-
-int main(int argc, char **argv) 
-{
-  rclcpp::init(argc, argv);
-
-  auto node = std::make_shared<j2735_convertor::Node>(rclcpp::NodeOptions());
-  
-  rclcpp::executors::MultiThreadedExecutor executor;
-  executor.add_node(node->get_node_base_interface());
-  executor.spin();
-
-  rclcpp::shutdown();
-
-  return 0;
-}
